requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.1.3
blinker==1.7.0
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.2.0
customtkinter==5.2.0
cycler==0.12.1
darkdetect==0.8.0
distlib==0.3.8
filelock==3.15.4
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
fonttools==4.47.2
frozendict==2.4.0
frozenlist==1.4.1
html5lib==1.1
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonify==0.5
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.3
multidict==6.0.5
multitasking==0.0.11
musicbrainzngs==0.7.1
packaging==23.2
pbkdf2==1.3
peewee==3.17.0
pipenv==2024.0.1
platformdirs==4.2.2
prettytable==3.5.0
psycopg2==2.9.9
pyaes==1.6.1
PyJWT==2.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.8
requests==2.31.0
scipy==1.11.4
six==1.16.0
soupsieve==2.5
threadpoolctl==3.2.0
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.2
virtualenv==20.26.3
waitress==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.0.1
yarl==1.9.4
